SEQUOIADBMAINSTREAM-3133 使用内置SQL插入timestamp类型记录，1970年前的时间戳记录异常
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.8@34359 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bson/bson_db.h
@@ -40,8 +40,8 @@ namespace bson {
     @param time - in millis (but stored in seconds)
     */
     inline BSONObjBuilder& BSONObjBuilder::appendTimestamp(
-      const StringData& fieldName, unsigned long long time, unsigned int inc) {
-        OpTime t( (unsigned) (time / 1000) , inc );
+      const StringData& fieldName, long long time, unsigned int inc) {
+        OpTime t( (time / 1000) , inc );
         appendTimestamp( fieldName , t.asDate() );
         return *this;
     }
